requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
djangorestframework==3.12.2
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.10
pkg-resources==0.0.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
websocket-client==0.48.0
